+++ src/main/scala/com/isgneuro/zeppelinotl/Result.scala
@@ -1,6 +1,0 @@
-package com.isgneuro.zeppelinotl
-
-import com.isgneuro.otp.connector.Dataset
-import org.apache.zeppelin.interpreter.InterpreterResult
-
-case class Result(interpreterResult: InterpreterResult, dataset: Option[Dataset])