+++ src/test/scala/com/isgneuro/zeppelinotl/OTLInterpreterTest.scala
@@ -24,49 +24,30 @@ class OTLInterpreterTest extends FlatSpec with Matchers {
 
   val ctx: InterpreterContext = new InterpreterContext.Builder().build
 
-  "OTLInterpreter" should "fail if incorrect credentials specified in propeties" ignore {
-    properties.setProperty("OTP.rest.auth.password", "wrongPass")
-    val query = "| makeresults"
-    val interp = new OTLInterpreter(properties)
-    val res = interp.interpret(query, ctx)
-    res.code().toString should be("ERROR")
-    res.toString should include("Fail to authenticate")
-    properties.setProperty("OTP.rest.auth.password", "12345678")
-  }
+  it should "parse json lines from dataset and convert them to appropriate format" in {
 
-  it should "convert 'earliest' and 'latest' to unix time" in {
-    val query = "| getdata object=devices_h  earliest =  2019/01/01:0:0:0 latest= 2020/01/01:0:0:0 | stats min(_time) as mint"
-    val expected = "| getdata object=devices_h  earliest =  1546290000 latest= 1577826000 | stats min(_time) as mint"
-    Query(query).convertTimeRange.query should be(expected)
-  }
+    //otlquery - "makeresults count=2 | eval x=2 | eval y=x*x | fields x, y"
 
-  it should "not convert 'earliest' and 'latest' to unix time if they are not in the first command" in {
-    val query = """makeresults | eval f1 = "earliest = 2019/01/01:0:0:0" | eval f2 = "latest=2020/01/01:0:0:0" """
-    val expected = query
-    Query(query).convertTimeRange.query should be(expected)
-  }
+    val datasetJsonLines = Seq("""{"x":2,"y":4}""", """{"x":2,"y":4}""")
+    val datasetSchema = """`x` INT,`y` INT"""
 
-  it should "extract time range from query if specified" in {
-    val query = "| getdata object=devices_h  earliest = 2019/01/01:0:0:0  latest  =2020/01/01:0:0:0 | stats min(_time) as mint"
-    Query(query).earliest should be(1546290000)
-    Query(query).latest should be(1577826000)
-  }
-
-  ignore should "parse json lines from dataset and convert them to appropriate format" in {
-    val query = "makeresults count=2 | eval x=2 | eval y=x*x | fields x, y"
-    val connInfo = ConnectionInfo(properties.getProperty("OTP.rest.host"), properties.getProperty("OTP.rest.port"), ssl = false)
-    val conn = Connector(connInfo, properties.getProperty("OTP.rest.auth.username"), properties.getProperty("OTP.rest.auth.password"))
-    val dataset = conn.jobs.create(query).getDataset
     val expected = "x\ty\n2\t4\n2\t4"
     val interp = new OTLInterpreter(properties)
-    interp.parseEvents(dataset.jsonLines, dataset.schema) should be(expected)
+    interp.parseEvents(datasetJsonLines, datasetSchema) should be(expected)
   }
 
-  ignore should "convert List to correct string without 'List' word in resulting dataset" in {
-    val query = "makeresults count=2 | streamstats count as x | stats list(x) as xs"
-    val interp = new OTLInterpreter(properties)
-    val res = interp.interpret(query, ctx)
-    res.toJson should include(""""data":"xs\n(1, 2)"""")
+  it should "convert List to correct string without 'List' word in resulting dataset" in {
+
+    //"makeresults count=2 | streamstats count as x | stats list(x) as xs"
+
+    val datasetJsonLines = Seq("""{"xs":[1,2]}""")
+    val datasetSchema = """`xs` ARRAY<BIGINT>"""
+
+    val interpreter = new OTLInterpreter(properties)
+    val result = interpreter.parseEvents(datasetJsonLines, datasetSchema)
+    val expected = "xs\n(1, 2)"
+
+    result should be(expected)
   }
 
   ignore should "get query as string, run Connector, get and parse resulting dataset" in {
@@ -103,44 +84,24 @@ class OTLInterpreterTest extends FlatSpec with Matchers {
     res.toString should include("'where' command")
   }
 
-  ignore should "convert _time column to millisecs in resulting DF according to OTP.query.convertTime property" in {
-    val query = "| makeresults "
+  it should "convert _time column to millisecs in resulting DF according to OTP.query.convertTime property" in {
+
+    val datasetJsonLines = Seq("""{"_time":1643889830}""")
+    val datasetSchema = """`_time` BIGINT"""
+
     properties.setProperty("OTP.query.convertTime", "true")
-    val interpTrue = new OTLInterpreter(properties)
-    val dsTrue = interpTrue.interpret(query, ctx)
-    val resultValueTrue = dsTrue.toString.split("\n").last
-    resultValueTrue.toLong % 1000 should be(0)
-    resultValueTrue should have length 13
+
+    val interpretResult = new OTLInterpreter(properties).parseEvents(datasetJsonLines, datasetSchema)
 
     properties.setProperty("OTP.query.convertTime", "false")
-    val interpFalse = new OTLInterpreter(properties)
-    val dsFalse = interpFalse.interpret(query, ctx)
-    val resultValueFalse = dsFalse.toString.split("\n").last
-    resultValueFalse should have length 10
-  }
 
-  it should "extract cache ID from query" in {
-    val query = "| makeresults | zput  id  =  test123  | eval a=1"
-    val cleanedQuery = Query(query).getCacheId
-    cleanedQuery.query should be("| makeresults | eval a=1")
-    cleanedQuery.cacheId should be('defined)
-    cleanedQuery.cacheId.get should be("test123")
-  }
+    val interpretResult2 = new OTLInterpreter(properties).parseEvents(datasetJsonLines, datasetSchema)
 
-  it should "return the same query if no cacheId specified in the end of query" in {
-    val query = "| makeresults | eval a=1"
-    val cleanedQuery = Query(query).getCacheId
-    cleanedQuery.query should be(query)
-    cleanedQuery.cacheId should be(None)
-  }
+    val expected = "_time\n1643889830000"
+    val expected2 = "_time\n1643889830"
 
-  ignore should "put dataset to cache if cacheId specified in query" in {
-    val query = "| makeresults count=2 | eval a=1 | fields a | zput  id =      test1"
-    val ctxRP = InterpreterContextHelper.setResourcePool(ctx)
-    val interp = new OTLInterpreter(properties)
-    val _ = interp.interpret(query, ctxRP)
-    val expected = """[{"a":1}, {"a":1}]"""
-    ctxRP.getResourcePool.get("test1").get.toString should be(expected)
+    interpretResult should be(expected)
+    interpretResult2 should be(expected2)
   }
 
   it should "get tokens' values from resource pool and put them into query" in {