+++ ReleaseNotes.md
@@ -1,3 +1,10 @@
+# [1.1.0] - Remove **earliest**, **latest** functions
+### Changed
+- project structure
+- some tests
+### Removed
+- **earliest**, **latest** functions support
+
 # [1.0.0] - Refactor and functional change
 
 ### Changed